googleslides/requirements.txt
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.147.0
